% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aot.R
\name{ph_aot}
\alias{ph_aot}
\title{phylomatic}
\usage{
ph_aot(traits = NULL, traits_file = NULL, phylo = NULL,
  phylo_file = NULL, randomizations = 999, trait_contrasts = 1,
  ebl_unstconst = FALSE)
}
\arguments{
\item{traits}{(data.frame) trait data.frame}

\item{traits_file}{(character) path to traits file}

\item{phylo}{(character) phylogeny as a newick string, will be written to
a temp file if provided}

\item{phylo_file}{(character) path to phylo file}

\item{randomizations}{(numeric) number of randomizations. Default: 999}

\item{trait_contrasts}{(numeric) Specify which trait should be used as 'x'
variable for contrasts. Default: 1}

\item{ebl_unstconst}{(logical) Use equal branch lengths and unstandardized
contrasts. Default: \code{FALSE}}
}
\description{
phylomatic
}
\examples{
traits_file <- system.file("examples/traits_aot", package = "phylocom")
phylo_file <- system.file("examples/phylo_aot", package = "phylocom")

# from data.frame
traitsdf_file <- system.file("examples/traits_aot_df", package = "phylocom")
traits <- read.table(text = readLines(traitsdf_file), header = TRUE,
  stringsAsFactors = FALSE)
phylo_str <- readLines(phylo_file)
(res <- ph_aot(traits = traits, phylo = phylo_str))

# from files
traits_file2 <- tempfile()
cat(traits_str, file = traits_file2, sep = '\\n')
phylo_file2 <- tempfile()
cat(phylo_str, file = phylo_file2, sep = '\\n')
(res <- ph_aot(traits_file = traits_file2, phylo_file = phylo_file2))
}

