% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bladj.R
\name{ph_bladj}
\alias{ph_bladj}
\title{bladj}
\usage{
ph_bladj(ages = NULL, ages_file = NULL, phylo = NULL, phylo_file = NULL)
}
\arguments{
\item{ages}{(data.frame) ages data.frame}

\item{ages_file}{(character) path to traits file}

\item{phylo}{(character) phylogeny as a newick string, will be written to
a temp file if provided}

\item{phylo_file}{(character) path to phylo file}
}
\value{
newick string
}
\description{
bladj
}
\examples{
library(phytools)

ages_file <- system.file("examples/ages", package = "phylocom")
phylo_file <- system.file("examples/phylo_bladj", package = "phylocom")

# from data.frame
ages_df <- data.frame(
  a = c('malpighiales','salicaceae','fabaceae','rosales','oleaceae',
        'gentianales','apocynaceae','rubiaceae'),
  b = c(81,20,56,76,47,71,18,56)
)
phylo_str <- readLines(phylo_file)
(res <- ph_bladj(ages = ages_df, phylo = phylo_str))
plot(phytools::read.newick(text = res))

# from files
ages_file2 <- file.path(tempdir(), "ages")
write.table(ages_df, file = ages_file2, row.names = FALSE,
  col.names = FALSE, quote = FALSE)
phylo_file2 <- tempfile()
cat(phylo_str, file = phylo_file2, sep = '\\n')
(res <- ph_bladj(ages_file = ages_file2, phylo_file = phylo_file2))
plot(phytools::read.newick(text = res))
}

