% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrait.R
\name{ph_comtrait}
\alias{ph_comtrait}
\title{comtrait}
\usage{
ph_comtrait(sample, traits, metric = "variance", null_model = 0,
  randomizations = 999, abundance = TRUE)
}
\arguments{
\item{sample}{(data.frame/character) sample data.frame or path to a
sample file}

\item{traits}{(data.frame/character) traits data.frame or path to a
traits file}

\item{metric}{(integer) metric to calculate. One of variance, mpd, mntd,
or range (converted to phylocom integer format internally)}

\item{null_model}{(integer) which null model to use. See Details.}

\item{randomizations}{(numeric) number of randomizations. Default: 999}

\item{abundance}{(logical) If \code{TRUE} (default) computed accounting
for abundance. Otherwise, uses presence-absence.}
}
\value{
data.frame
}
\description{
comtrait
}
\section{Null models}{

\itemize{
 \item 0 - This null model shuffles trait values across species.
 \item 1 - Species in each sample become random draws from sample pool.
 This null model maintains the species richness of each sample, but
 the identities of the species occurring in each sample are randomized.
 For each sample, species are drawn without replacement from the list of
 all species actually occurring in at least one sample
 \item 2 - Species in each sample become random draws from traits data.
 This null model maintains the species richness of each sample, but the
 identities of the species occurring in each sample are randomized. For
 each sample, species are drawn without replacement from the list of all
 species with trait values. This function is redundant since by definition
 the sample and trait species must match, but is included for consistency
 with the comstruct function.
 \item 3 - Independent swap: Same as for \code{\link{ph_comdist}} and
 \code{\link{ph_comstruct}}
}
}
\examples{
sfile <- system.file("examples/sample_comstruct", package = "phylocomr")
tfile <- system.file("examples/traits_aot", package = "phylocomr")

# from files
sample_str <- paste0(readLines(sfile), collapse = "\\n")
sfile2 <- tempfile()
cat(sample_str, file = sfile2, sep = '\\n')

traits_str <- paste0(readLines(tfile), collapse = "\\n")
tfile2 <- tempfile()
cat(traits_str, file = tfile2, sep = '\\n')

ph_comtrait(sample = sfile2, traits = tfile2)


# from data.frame
sampledf <- read.table(sfile, header = FALSE,
  stringsAsFactors = FALSE)
traitsdf_file <- system.file("examples/traits_aot_df",
  package = "phylocomr")
traitsdf <- read.table(text = readLines(traitsdf_file), header = TRUE,
  stringsAsFactors = FALSE)
# ph_comtrait(sample = sampledf, traits = traitsdf)
}

