% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executables.R
\name{executables}
\alias{ecovolve}
\alias{executables}
\alias{phylocom}
\alias{phylomatic}
\title{Executables}
\usage{
ecovolve(args = "--help", stdout = "")

phylocom(args = "help", stdout = "")

phylomatic(args = "--help", stdout = "")
}
\description{
Executables
}
\examples{
ecovolve()
phylocom()
phylomatic()
}

