% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executables.R
\name{executables}
\alias{executables}
\alias{ecovolve}
\alias{phylocom}
\alias{phylomatic}
\title{Executables}
\usage{
ecovolve(args = "--help", intern = FALSE)

phylocom(args = "help", intern = FALSE)

phylomatic(args = "--help", intern = FALSE)
}
\arguments{
\item{args}{a character vector of arguments to command.}

\item{intern}{capture output as character vector. Default: \code{FALSE}}
}
\description{
Executables
}
\examples{
ecovolve()
phylocom()
phylomatic()
}
