% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomatic.R
\name{ph_phylomatic}
\alias{ph_phylomatic}
\title{phylomatic}
\usage{
ph_phylomatic(taxa, phylo, tabular = FALSE, lowercase = FALSE,
  nodes = FALSE)
}
\arguments{
\item{taxa}{(character) all taxa as a character vector (will be written to
a temp file if provided) - OR a path to taxa file. Required. See Details.}

\item{phylo}{(character/phylo) phylogeny as a \code{phylo} object, a newick
string (both will be written to a temp file if provided) - OR a path to a
file with the newick string. Required.}

\item{tabular}{(logical) Output a tabular representation of phylogeny.
Default: \code{FALSE}}

\item{lowercase}{(logical) Convert all chars in taxa file to lowercase.
Default: \code{FALSE}}

\item{nodes}{(logical) label all nodes with default names.
Default: \code{FALSE}}
}
\description{
phylomatic
}
\details{
The \code{taxa} character vector must have each element of the
form \code{family/genus/genus_epithet}. If a file is passed in, each
line should have a \code{family/genus/genus_epithet} string - make sure
only one per line, and a newline (i.e., press ENTER) at the end of
each line
}
\examples{
library(phytools)

taxa_file <- system.file("examples/taxa", package = "phylocomr")
phylo_file <- system.file("examples/phylo", package = "phylocomr")

# from strings
(taxa_str <- readLines(taxa_file))
(phylo_str <- readLines(phylo_file))
(tree <- ph_phylomatic(taxa = taxa_str, phylo = phylo_str))

# from files
taxa_file2 <- tempfile()
cat(taxa_str, file = taxa_file2, sep = '\\n')
phylo_file2 <- tempfile()
cat(phylo_str, file = phylo_file2, sep = '\\n')
(tree <- ph_phylomatic(taxa = taxa_file2, phylo = phylo_file2))
library(ape)
plot(read.newick(text = tree))
}

