% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomatic.R
\name{ph_phylomatic}
\alias{ph_phylomatic}
\title{phylomatic}
\usage{
ph_phylomatic(taxa = NULL, taxa_file = NULL, phylo = NULL,
  phylo_file = NULL, tabular = FALSE, lowercase = FALSE, nodes = FALSE)
}
\arguments{
\item{taxa}{(character) all taxa as a character vector, will be written to
a temp file if provided}

\item{taxa_file}{(character) path to taxa file}

\item{phylo}{(character) phylogeny as a newick string, will be written to
a temp file if provided}

\item{phylo_file}{(character) path to phylo file}

\item{tabular}{(logical) Output a tabular representation of phylogeny.
Default: \code{FALSE}}

\item{lowercase}{(logical) Convert all chars in taxa file to lowercase.
Default: \code{FALSE}}

\item{nodes}{(logical) label all nodes with default names.
Default: \code{FALSE}}
}
\description{
phylomatic
}
\examples{
taxa_file <- system.file("examples/taxa", package = "phylocom")
phylo_file <- system.file("examples/phylo", package = "phylocom")

# from strings
(taxa_str <- readLines(taxa_file))
(phylo_str <- readLines(phylo_file))
(tree <- ph_phylomatic(taxa = taxa_str, phylo = phylo_str))

# from files
taxa_file2 <- tempfile()
cat(taxa_str, file = taxa_file2, sep = '\\n')
phylo_file2 <- tempfile()
cat(phylo_str, file = phylo_file2, sep = '\\n')
(tree <- ph_phylomatic(taxa_file = taxa_file2, phylo_file = phylo_file2))
library(ape)
plot(read.tree(text = tree))

}

